\name{cubic}
\alias{cubic}
\docType{data}
\title{A Data Set for the Example of the Qprog Function}
\description{This data set is used for the example of the qprog function.  }
\usage{data(cubic)}
\format{A data frame with 50 observations on the following 2 variables.
\describe{
\item{\code{x}}{The predictor vector.}
\item{\code{y}}{The response vector.}  }  }
\details{We use the qprog function to fit a constrained cubic to this data set. The constraint is that the true regression is increasing, convex and nonnegative. }
\source{STAT640 HW 14 given by Dr. Meyer.  }

\keyword{data sets}
