\name{decr}

\alias{decr}

\title{Specify a Decreasing Shape-Restriction in a SHAPEREG Formula}

\description{
A symbolic routine to define that the mean vector is decreasing in a predictor in a formula argument to shapereg.}

\usage{decr(x, numknots = 0, knots = 0, space = "E")}

\arguments{
  \item{x}{A numeric predictor which has the same length as the response vector.}
   \item{numknots}{The number of knots used to smoothly constrain a predictor. The value should be \eqn{0} for a shape-restricted predictor without smoothing. The default value is \eqn{0}.}
  \item{knots}{The knots used to smoothly constrain a predictor. The value should be \eqn{0} for a shape-restricted predictor without smoothing. The default value is \eqn{0}.}
 \item{space}{A character specifying the method to create knots. It will not be used for a shape-restricted predictor without smoothing. The default value is "E".}
}

\value{The vector x with the shape attribute, i.e., shape: 2 ("decreasing").}

\details{
  "decr" returns the vector "x" and imposes on it two attributes: name and shape. 

  The shape attribute is 2 ("decreasing"), and according to the value of the vector itself and this attribute, the cone edges of the cone generated by the constraint matrix, which constrains the relationship between the mean vector and "x" to be decreasing,  will be made. The cone edges are a set of basis employed in the hinge algorithm. 

  Note that "decr" does not make the corresponding cone edges itself. It sets things up to a subroutine called makedelta in coneproj.

  See references cited in this section for more details.
}

\references{
  Meyer, M. C. (2013b) A simple new algorithm for quadratic programming with applications in statistics. \emph{Communications in Statistics \bold{42(5)}}, 1126--1139. 
}

\author{
  Mary C. Meyer and Xiyue Liao
}

\examples{
  data(cubic)

  # extract x
  x <- - cubic$x

  # extract y
  y <- cubic$y

  # regress y on x with the shape restriction: "decreasing"
  ans <- shapereg(y ~ decr(x))

  # make a plot
  par(mar = c(4, 4, 1, 1))
  plot(x, y, cex = .7, xlab = "x", ylab = "y")
  lines(x, fitted(ans), col = 2)
  legend("topleft", bty = "n", "shapereg: decreasing fit", col = 2, lty = 1)
}

\seealso{
  \code{\link{decr.conc}}, \code{\link{decr.conv}}
}

\keyword{shape routine}
