\name{FEV}
\alias{FEV}
\docType{data}
\title{Forced Expiratory Volume}
\description{This data set consists of 654 observations on children aged 3 to 19. Forced Expiratory Volume (FEV), which is a measure of lung capacity, is the variable in interest. Age and height are two continuous predictors. Sex and smoke are two categorical predictors.

}
\usage{data(FEV)}
\format{
  A data frame with 654 observations on the following 5 variables.
  \describe{
    \item{\code{age}}{Age of the 654 children.}
    \item{\code{FEV}}{Forced expiratory volume(liters).}
    \item{\code{height}}{Height(inches).}
    \item{\code{sex}}{Female is 0. Male is 1.}
    \item{\code{smoke}}{Nonsmoker is 0. Smoker is 1.}
  }
}

\source{
Rosner, B. (1999) \emph{Fundamentals of Biostatistics, 5th Ed., Pacific Grove, CA: Duxbur}.

Michael J. Kahn (2005) An Exhalent Problem for Teaching Statistics \emph{Journal of Statistics Education \bold{Volume 13, Number 2}}.
}

\keyword{data sets}
